fix: move missed lambda users-ms to typescript
This module was missed during update
