+++ services/users-ms/src/lambdas/findUser.js
@@ -1,10 +1,10 @@
 /* eslint-disable no-console */
 import to from 'await-to-js';
 
-import config from '../../../config';
-import * as dynamoDb from '../../../libs/dynamoDb';
+import config from '../libs/config';
+import * as dynamoDb from '../libs/dynamoDb';
 import putUserEvent from '../helpers/putUserEvent';
-import log from '../../../libs/logs';
+import log from '../libs/logs';
 
 export async function main(event, context) {
   const userDetail = event.detail;
