+++ services/users-ms/src/libs
@@ -1,0 +1,1 @@
+../../../libs
